+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/v4/src/webkit_webview_platform.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:webview_flutter_platform_interface/v4/webview_flutter_platform_interface.dart';
+import 'package:webview_flutter_wkwebview/src/v4/src/webkit_webview_cookie_manager.dart';
 
 import 'webkit_navigation_delegate.dart';
 import 'webkit_webview_controller.dart';
@@ -29,5 +30,12 @@ class WebKitWebViewPlatform extends WebViewPlatform {
     PlatformWebViewWidgetCreationParams params,
   ) {
     return WebKitWebViewWidget(params);
+  }
+
+  @override
+  WebKitWebViewCookieManager createPlatformCookieManager(
+    PlatformWebViewCookieManagerCreationParams params,
+  ) {
+    return WebKitWebViewCookieManager(params);
   }
 }