+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/v4/webview_flutter_wkwebview.dart
@@ -6,6 +6,6 @@ library webview_flutter_wkwebview;
 
 export 'src/webkit_navigation_delegate.dart';
 export 'src/webkit_webview_controller.dart';
-export 'src/webkit_webview_controller.dart';
+export 'src/webkit_webview_cookie_manager.dart';
 export 'src/webkit_webview_platform.dart';
 export 'src/webkit_webview_widget.dart';