+++ packages/shared_preferences/shared_preferences_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum Flutter version to 2.10.
+
 ## 2.0.12
 
 * Fixes library_private_types_in_public_api, sort_child_properties_last and use_key_in_widget_constructors