+++ packages/url_launcher/url_launcher_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum Flutter version to 2.10.
+
 ## 2.1.0
 
 * Adds a new `launchUrl` method corresponding to the new app-facing interface.